Cloudserver features (#1)
* Add CloudServer endpoints

* Update README.md

* Add error id to NitrapiErrorException for support requests

* Fix warnings

* Fix event logs

* Add service comment

* Add Cloud Server firewall

* Fix pricing calculation

* Add an exception when the access token is invalid

* Round advice

* Don't show services that we can not load due to errors

* Fix extending dynamic cloud servers

* Add apps and systemd

* Add support authorization

* Cancel and delete services

* Add ip binding

* Get single systemd unit

* Add journal endpoint

* Get list of all users on cloud server

* Add access token info

* Add daemon fields for image

* Fix firewall and app manager

+++ Pod/Classes/order/PartRentalOption.swift
@@ -20,7 +20,8 @@ open class PartRentalOption: Mappable {
 
 open class Price: Mappable {
     open fileprivate(set) var count: Int!
-    open fileprivate(set) var price: Int!
+    /// Price in cents. Should be a whole number unless you deal with dynamic cloud servers.
+    open fileprivate(set) var price: Double!
